+++ frameworks/qwik/src/types/form.ts
@@ -23,6 +23,10 @@ export interface FormStore<TSchema extends FormSchema = FormSchema>
    */
   readonly isTouched: ReadonlySignal<boolean>;
   /**
+   * Whether any field in the form has been edited.
+   */
+  readonly isEdited: ReadonlySignal<boolean>;
+  /**
    * Whether any field in the form differs from its initial value.
    */
   readonly isDirty: ReadonlySignal<boolean>;