+++ frameworks/svelte/src/runes/useField/useField.svelte.ts
@@ -56,6 +56,7 @@ export function useField(
 
   const input = $derived(getFieldInput(internalFieldStore));
   const isTouched = $derived(getFieldBool(internalFieldStore, 'isTouched'));
+  const isEdited = $derived(getFieldBool(internalFieldStore, 'isEdited'));
   const isDirty = $derived(getFieldBool(internalFieldStore, 'isDirty'));
   const isValid = $derived(!getFieldBool(internalFieldStore, 'errors'));
 
@@ -71,6 +72,9 @@ export function useField(
     },
     get isTouched() {
       return isTouched;
+    },
+    get isEdited() {
+      return isEdited;
     },
     get isDirty() {
       return isDirty;