+++ frameworks/vue/src/composables/useFieldArray/useFieldArray.ts
@@ -58,6 +58,9 @@ export function useFieldArray(
   const isTouched = computed(() =>
     getFieldBool(internalFieldStore.value, 'isTouched')
   );
+  const isEdited = computed(() =>
+    getFieldBool(internalFieldStore.value, 'isEdited')
+  );
   const isDirty = computed(() =>
     getFieldBool(internalFieldStore.value, 'isDirty')
   );
@@ -77,6 +80,9 @@ export function useFieldArray(
     },
     get isTouched() {
       return isTouched.value;
+    },
+    get isEdited() {
+      return isEdited.value;
     },
     get isDirty() {
       return isDirty.value;