+++ packages/methods/src/remove/remove.ts
@@ -70,8 +70,9 @@ export function remove<
         );
       }
 
-      // Mark field array as touched and update dirty state
+      // Mark field array as touched and edited and update dirty state
       internalArrayStore.isTouched.value = true;
+      internalArrayStore.isEdited.value = true;
       internalArrayStore.isDirty.value =
         internalArrayStore.startItems.value.join() !== newItems.join();