+++ packages/methods/src/reset/reset.ts
@@ -28,6 +28,10 @@ interface ResetBaseConfig {
    */
   readonly keepTouched?: boolean | undefined;
   /**
+   * Whether to keep the edited state during reset. Defaults to false.
+   */
+  readonly keepEdited?: boolean | undefined;
+  /**
    * Whether to keep the error messages during reset. Defaults to false.
    */
   readonly keepErrors?: boolean | undefined;
@@ -136,6 +140,11 @@ export function reset(
         // Reset is touched if it is not to be kept
         if (!config?.keepTouched) {
           internalFieldStore.isTouched.value = false;
+        }
+
+        // Reset is edited if it is not to be kept
+        if (!config?.keepEdited) {
+          internalFieldStore.isEdited.value = false;
         }
 
         // Reset start input to initial input