+++ packages/core/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the library will be documented in this file.
 
+## vX.X.X (Month DD, YYYY)
+
+- Add `isEdited` field state that is set when a field's value changes, is not set on focus, and stays set after reverting to the initial value
+
 ## v0.8.0 (June 15, 2026)
 
 - Add `focusFieldElement` utility that focuses the first focusable element of a field store